requirements.txt
customtkinter>=5.2.0
cryptography>=42.0.0
argon2-cffi>=23.1.0
pyperclip>=1.8.2
